modularize cos, cosh, sin, sinh, tan, tanh
