+++ tfjs-core/src/ops/ops.ts
@@ -52,6 +52,8 @@ export {conv2d} from './conv2d';
 export {conv2dTranspose} from './conv2d_transpose';
 export {conv3d} from './conv3d';
 export {conv3dTranspose} from './conv3d_transpose';
+export {cos} from './cos';
+export {cosh} from './cosh';
 export {cumsum} from './cumsum';
 export {depthToSpace} from './depth_to_space';
 export {depthwiseConv2d} from './depthwise_conv2d';
@@ -123,6 +125,8 @@ export {reverse4d} from './reverse_4d';
 export {selu} from './selu';
 export {separableConv2d} from './separable_conv2d';
 export {sign} from './sign';
+export {sin} from './sin';
+export {sinh} from './sinh';
 export {spaceToBatchND} from './space_to_batch_nd';
 export {split} from './split';
 export {square} from './square';
@@ -131,6 +135,8 @@ export {squeeze} from './squeeze';
 export {stack} from './stack';
 export {sub} from './sub';
 export {sum} from './sum';
+export {tan} from './tan';
+export {tanh} from './tanh';
 export {tile} from './tile';
 export {truncatedNormal} from './truncated_normal';
 export {unsortedSegmentSum} from './unsorted_segment_sum';
