+++ tfjs-core/src/ops/tanh.ts
@@ -1,0 +1,49 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {ENGINE} from '../engine';
+import {Tanh, TanhInputs} from '../kernel_names';
+import {Tensor} from '../tensor';
+import {NamedTensorMap} from '../tensor_types';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+
+import {op} from './operation';
+
+/**
+ * Computes hyperbolic tangent of the input `tf.Tensor` element-wise: `tanh(x)`
+ *
+ * ```js
+ * const x = tf.tensor1d([0, 1, -1, 70]);
+ *
+ * x.tanh().print();  // or tf.tanh(x)
+ * ```
+ * @param x The input tensor.
+ */
+/** @doc {heading: 'Operations', subheading: 'Basic math'} */
+function tanh_<T extends Tensor>(x: T|TensorLike): T {
+  const $x = convertToTensor(x, 'x', 'tanh');
+
+  const inputs: TanhInputs = {x: $x};
+
+  return ENGINE.runKernelFunc((backend, save) => {
+    const y = backend.tanh($x);
+    save([y]);
+    return y;
+  }, inputs as {} as NamedTensorMap, null /* grad */, Tanh);
+}
+export const tanh = op({tanh_});
