+++ tfjs-core/src/ops/unary_ops_test.ts
@@ -17,8 +17,7 @@
 
 import * as tf from '../index';
 import {ALL_ENVS, describeWithFlags} from '../jasmine_util';
-import {expectArraysClose, TEST_EPSILON_FLOAT16} from '../test_util';
-import * as util from '../util';
+import {expectArraysClose} from '../test_util';
 
 describeWithFlags('step', ALL_ENVS, () => {
   it('with 1d tensor', async () => {
@@ -1340,599 +1339,6 @@ describeWithFlags('expm1', ALL_ENVS, () => {
   it('throws for string tensor', () => {
     expect(() => tf.expm1('q'))
         .toThrowError(/Argument 'x' passed to 'expm1' must be numeric/);
-  });
-});
-
-describeWithFlags('sin', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.sin(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.sin(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.sin(a);
-    expectArraysClose(await res.data(), [Math.sin(4), NaN, Math.sin(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cos(5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sin(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cos(5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1, 2, 3, -5]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const gradients = tf.grad(a => tf.sin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [1 * Math.cos(-1), 2 * Math.cos(2), 3 * Math.cos(3), 4 * Math.cos(-5)],
-        1e-1);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.sin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [1 * Math.cos(-3), 2 * Math.cos(1), 3 * Math.cos(2), 4 * Math.cos(3)],
-        1e-1);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.sin({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'sin' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.sin(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.sin(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.sin('q'))
-        .toThrowError(/Argument 'x' passed to 'sin' must be numeric/);
-  });
-});
-
-describeWithFlags('cos', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.cos(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.cos(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.cos(a);
-    expectArraysClose(await res.data(), [Math.cos(4), NaN, Math.cos(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sin(5) * -1]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cos(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sin(5) * -1]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1, 2, 3, -5]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const gradients = tf.grad(a => tf.cos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [
-          1 * Math.sin(-1) * -1, 2 * Math.sin(2) * -1, 3 * Math.sin(3) * -1,
-          4 * Math.sin(-5) * -1
-        ],
-        1e-1);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.cos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [
-          1 * Math.sin(-3) * -1, 2 * Math.sin(1) * -1, 3 * Math.sin(2) * -1,
-          4 * Math.sin(3) * -1
-        ],
-        1e-1);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.cos({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'cos' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.cos(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.cos(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.cos('q'))
-        .toThrowError(/Argument 'x' passed to 'cos' must be numeric/);
-  });
-});
-
-describeWithFlags('tan', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.tan(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.tan(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.tan(a);
-    expectArraysClose(await res.data(), [Math.tan(4), NaN, Math.tan(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tan(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 / (Math.cos(0.5) * Math.cos(0.5))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tan(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 / (Math.cos(0.5) * Math.cos(0.5))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.tan(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (Math.cos(aValues[i]) * Math.cos(aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    // The grad(tan(x)) which relies on 1/cos(x) is less precise on Windows.
-    expectArraysClose(await gradients.data(), expected, TEST_EPSILON_FLOAT16);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.tan(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (Math.cos(aValues[i]) * Math.cos(aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.tan({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'tan' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.tan(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.tan(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.tan('q'))
-        .toThrowError(/Argument 'x' passed to 'tan' must be numeric/);
-  });
-});
-
-describeWithFlags('sinh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.sinh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.sinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.sinh(a);
-    expectArraysClose(await res.data(), [Math.sinh(4), NaN, Math.sinh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sinh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cosh(0.5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sinh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cosh(0.5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.sinh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.cosh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.sinh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.cosh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.sinh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'sinh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const result = tf.sinh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.sinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.sinh('q'))
-        .toThrowError(/Argument 'x' passed to 'sinh' must be numeric/);
-  });
-});
-
-describeWithFlags('cosh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.cosh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.cosh(values[i]);
-    }
-
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.cosh(a);
-    expectArraysClose(await res.data(), [Math.cosh(4), NaN, Math.cosh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cosh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sinh(0.5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cosh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sinh(0.5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.cosh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.sinh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.cosh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.sinh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.cosh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'cosh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const result = tf.cosh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.cosh(values[i]);
-    }
-
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.cosh('q'))
-        .toThrowError(/Argument 'x' passed to 'cosh' must be numeric/);
-  });
-});
-
-describeWithFlags('tanh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.tanh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = util.tanh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.tanh(a);
-    expectArraysClose(await res.data(), [util.tanh(4), NaN, util.tanh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tanh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 * (1 - (Math.tanh(0.5) * Math.tanh(0.5)))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tanh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 * (1 - (Math.tanh(0.5) * Math.tanh(0.5)))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.tanh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] =
-          dyValues[i] * (1 - (Math.tanh(aValues[i]) * Math.tanh(aValues[i])));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.tanh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] =
-          dyValues[i] * (1 - (Math.tanh(aValues[i]) * Math.tanh(aValues[i])));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.tanh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'tanh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.tanh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = util.tanh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.tanh('q'))
-        .toThrowError(/Argument 'x' passed to 'tanh' must be numeric/);
   });
 });
