+++ tfjs-core/src/tests.ts
@@ -81,6 +81,8 @@ import './ops/conv2d_transpose_test';
 import './ops/conv3d_test';
 import './ops/conv3d_transpose_test';
 import './ops/conv_util_test';
+import './ops/cos_test';
+import './ops/cosh_test';
 import './ops/cosine_distance_test';
 import './ops/crop_and_resize_test';
 import './ops/cumsum_test';
@@ -154,6 +156,8 @@ import './ops/selu_test';
 import './ops/sigmoid_cross_entropy_test';
 import './ops/sign_test';
 import './ops/signal_ops_test';
+import './ops/sin_test';
+import './ops/sinh_test';
 import './ops/slice_test';
 import './ops/slice_util_test';
 import './ops/softmax_cross_entropy_test';
@@ -165,12 +169,13 @@ import './ops/stack_test';
 import './ops/strided_slice_test';
 import './ops/sub_test';
 import './ops/sum_test';
+import './ops/tan_test';
+import './ops/tanh_test';
 import './ops/tile_test';
 import './ops/topk_test';
 import './ops/transpose_test';
 import './ops/truncated_normal_test';
 import './ops/unary_ops_test';
-import './ops/undefined_test';
 import './ops/unsorted_segment_sum_test';
 import './ops/unstack_test';
 import './ops/where_async_test';
